requirements.txt
asgiref==3.7.2
autopep8==2.0.4
backports.zoneinfo==0.2.1
black==23.12.0
click==8.1.7
colorama==0.4.6
Django==4.2.8
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
isort==5.13.2
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
psycopg2==2.9.9
pycodestyle==2.11.1
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
